Tilemap & Overlay: fix too small data size for tile map index (u8 -> u32) and downstream sizes for it, tidy up buffer freeing, fix tile id overlay var size and move some local functions to static

+++ src/tilemap_overlay.h
@@ -15,7 +15,7 @@ void tilemap_overlay_setparams(uint8_t * p_overlaybuf_new,
 
 void tilemap_overlay_set_enables(int grid_enabled, int tilenums_enabled);
 
-void tilemap_overlay_apply(uint32_t map_size, uint8_t * map_tilelist);
+void tilemap_overlay_apply(uint32_t map_size, uint32_t * map_tilelist);
 
 void overlay_redraw_invalidate(void);
 void overlay_redraw_clear_flag(void);
